Move test stubs for Settings to entities test file
